replace p10k with oh-my-posh

+++ rc/myset.sh
@@ -8,25 +8,31 @@ source $USER_CUSTOM/rc/aliases.sh
 
 ####################################################
 # nnn
-export GUI=1
-export NNN_OPTS="aceEruUx"
-export NNN_COLORS='1234'
-export NNN_OPENER=$HOME/.config/nnn/plugins/nuke
-export NNN_COPIER=$HOME/.config/nnn/plugins/.cbcp
-export NNN_PLUG='e:suedit;'
-export NNN_SSHFS='sshfs -C -o delay_connect,idmap=user,follow_symlinks,auto_cache,cache_timeout=3600'
-export NNN_BMS="M:$HOME/.config/nnn/mounts/;P:$HOME/.config/nnn/plugins;D:$HOME/Downloads/;H:$HOME/"
+if type nnn > /dev/null; then
+    if [ -n "$DISPLAY" ]; then
+        export GUI=1
+    else
+        export GUI=0
+    fi
+    export NNN_OPTS="aceEuUx"
+    export NNN_COLORS='1234'
+    export NNN_OPENER=$HOME/.config/nnn/plugins/nuke
+    export NNN_COPIER=$HOME/.config/nnn/plugins/.cbcp
+    export NNN_PLUG='e:suedit;'
+    export NNN_SSHFS='sshfs -C -o delay_connect,idmap=user,follow_symlinks,auto_cache,cache_timeout=3600'
+    export NNN_BMS="M:$HOME/.config/nnn/mounts/;P:$HOME/.config/nnn/plugins;D:$HOME/Downloads/;H:$HOME/"
 
-n ()
-{
-    if [[ "${NNNLVL:-0}" -ge 1 ]]; then
-        echo "nnn is already running"
-        return
-    fi
-    NNN_TMPFILE="${XDG_CONFIG_HOME:-$HOME/.config}/nnn/.lastd"
-    \nnn "$@"
-    if [ -f "$NNN_TMPFILE" ]; then
-            . "$NNN_TMPFILE"
-            rm -f "$NNN_TMPFILE" > /dev/null
-    fi
-}
+    n ()
+    {
+        if [[ "${NNNLVL:-0}" -ge 1 ]]; then
+            echo "nnn is already running"
+            return
+        fi
+        NNN_TMPFILE="${XDG_CONFIG_HOME:-$HOME/.config}/nnn/.lastd"
+        \nnn "$@"
+        if [ -f "$NNN_TMPFILE" ]; then
+                . "$NNN_TMPFILE"
+                rm -f "$NNN_TMPFILE" > /dev/null
+        fi
+    }
+fi
